Fix: listの中身をlaundriesのindexに移動
